backend/HMS/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.4.0
PyJWT==2.11.0
sqlparse==0.5.5
